+++ python/aistore/sdk/get_batch/errors.py
@@ -1,0 +1,7 @@
+#
+# Copyright (c) 2025, NVIDIA CORPORATION. All rights reserved.
+#
+
+
+class MultipartDecodeError(Exception):
+    """Raised when multipart decoding fails."""